requirements.txt
discord-py==2.1.0
requests==2.28.2
bs4==0.0.1